fix for ATLAS_2017_I1598613, not sure what for anymore because HEPFORGE was down for too long
--HG--
branch : release-2-6-x

+++ analyses/pluginATLAS/ATLAS_2017_I1598613.cc
@@ -246,7 +246,7 @@ namespace Rivet {
   protected:
 
     /// Analysis-mode switch
-    int _mode;
+    size_t _mode;
 
     /// Histograms
     map<string, HistoHandler> _h;
